Fix no-any-return mypy errors with explicit type casts
Add explicit type casts to return values across 17 files to resolve
mypy no-any-return errors:
- float() for numpy/torch scalar returns
- int() for dict.get() and numpy returns
- bool() for numpy comparison returns
- str() for dynamic dispatch returns
- bytes() for async stream reads
- list() for numpy tolist() and dict.get() returns
- dict() for tokenizer returns
- cast() for complex types (FrameworkError, PolicyValueNetwork,
  MADueDiligenceConfig, type[BaseLLMClient], AsyncIterator[str])

+++ src/enterprise/use_cases/ma_due_diligence/agents.py
@@ -552,4 +552,4 @@ class ComplianceCheckAgent:
     def _compute_confidence(self, results: dict[str, Any]) -> float:
         """Compute confidence based on checks performed."""
         total_checks = sum(r.get("checks_performed", 0) for r in results.values())
-        return min(0.5 + total_checks * 0.05, 0.9)
+        return float(min(0.5 + total_checks * 0.05, 0.9))
